+++ frontends/qt/server/detectdarktheme_linux.go
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 //go:build linux
-// +build linux
 
 package main