+++ frontends/web/src/components/sidebar/sidebar.tsx
@@ -10,7 +10,7 @@ import { deregisterTest } from '@/api/keystores';
 import { getVersion } from '@/api/bitbox02';
 import { debug } from '@/utils/env';
 import { AppLogoInverted, Logo } from '@/components/icon/logo';
-import { CloseXWhite, Cog, CogGray, Coins, Device, Eject, Linechart, RedDot, ShieldGray } from '@/components/icon';
+import { CloseXWhite, CogGray, Coins, Device, Eject, Linechart, RedDot, ShieldGray } from '@/components/icon';
 import { getAccountsByKeystore } from '@/routes/account/utils';
 import { SkipForTesting } from '@/routes/device/components/skipfortesting';
 import { AppContext } from '@/contexts/AppContext';
@@ -177,9 +177,8 @@ const Sidebar = ({
             to={'/settings'}
             title={t('sidebar.settings')}
             onClick={handleSidebarItemClick}>
-            <div className="stacked">
+            <div className={style.single}>
               <CogGray alt={t('sidebar.settings')} />
-              <Cog alt={t('sidebar.settings')} />
             </div>
             <span className={style.sidebarLabel}>
               {t('sidebar.settings')}